+++ sparks_pif_converters/DSC/dsc_to_pif.py
@@ -6,6 +6,7 @@ import peakutils
 
 
 def netzsch_3500_to_pif(closed_csv):
+    print("FILE IDENTIFIED AS NETZSCH 3500 OUTPUT FILE: %s" % closed_csv)
 
     # create chemical system and property array
     my_pif = ChemicalSystem()
@@ -25,38 +26,39 @@ def netzsch_3500_to_pif(closed_csv):
         for index, row in enumerate(reader):
 
             # ensure row has values
-            if len(row) != 0:
+            if len(row) == 0:
+                continue
 
-                # set values based on row[0]
-                if '#IDENTITY:' in row[0]:
-                    my_pif.chemical_formula = row[1].strip()
+            # set values based on row[0]
+            if '#IDENTITY:' in row[0]:
+                my_pif.chemical_formula = row[1].strip()
 
-                if "#INSTRUMENT:" in row[0]:
-                    measurement_device = Instrument(name=row[1].split(" ")[0], model=row[1].split(" ")[2])
+            if "#INSTRUMENT:" in row[0]:
+                measurement_device = Instrument(name=row[1].split(" ")[0], model=row[1].split(" ")[2])
 
-                if "#SAMPLE MASS /mg:" in row[0]:
-                    prop = Property(name='sample mass', scalars=row[1], units='mg')
-                    my_pif.properties.append(prop)
+            if "#SAMPLE MASS /mg:" in row[0]:
+                prop = Property(name='sample mass', scalars=row[1], units='mg')
+                my_pif.properties.append(prop)
 
-                if "#DATE/TIME:" in row[0]:
-                    date = Value(name="Experiment date", scalars=row[1].strip())
+            if "#DATE/TIME:" in row[0]:
+                date = Value(name="Experiment date", scalars=row[1].strip())
 
-                if "#TYPE OF CRUCIBLE:" in row[0]:
-                    crucible = Value(name="Crucible", scalars=row[1].strip()+" "+row[2].strip())
+            if "#TYPE OF CRUCIBLE:" in row[0]:
+                crucible = Value(name="Crucible", scalars=row[1].strip()+" "+row[2].strip())
 
-                if "#PROTECTIVE GAS:" in row[0]:
-                    atmosphere = Value(name="Atmosphere", scalars=row[1].strip())
+            if "#PROTECTIVE GAS:" in row[0]:
+                atmosphere = Value(name="Atmosphere", scalars=row[1].strip())
 
-                # Temp indicates header row. Define header_row_index.
-                if '##Temp.' in row[0]:
-                    header_row_index = index
+            # Temp indicates header row. Define header_row_index.
+            if '##Temp.' in row[0]:
+                header_row_index = index
 
-                # Iterate through values after header_row.
-                if header_row_index:
-                    if index > header_row_index:
-                        temp_array.append(float(row[0]))
-                        time_array.append(float(row[1]))
-                        heat_capacity_array.append(float(row[2]))
+            # Iterate through values after header_row.
+            if header_row_index:
+                if index > header_row_index:
+                    temp_array.append(float(row[0]))
+                    time_array.append(float(row[1]))
+                    heat_capacity_array.append(float(row[2]))
 
     float_capacities = [float(h) for h in heat_capacity_array]
     peak_indexes = peakutils.peak.indexes(float_capacities, thres=0.35/max(float_capacities), min_dist=1)
@@ -78,7 +80,7 @@ def netzsch_3500_to_pif(closed_csv):
     my_pif.properties.append(heat_capacity)
 
     # print dump to check format
-    print (pif.dumps(my_pif, indent=4))
+    # print(pif.dumps(my_pif, indent=4))
 
     return my_pif
 
@@ -92,6 +94,6 @@ if __name__ == '__main__':
     for f in args.csv:
         print("PARSING: %s" % f)
         pifs = netzsch_3500_to_pif(f)
-
         f_out = f.replace(".csv", ".json")
+        print("OUTPUT FILE: %s" % f_out)
         pif.dump(pifs, open(f_out, "w"), indent=4)
